+++ includes/widgets/mashup-demo/data-provider-widget/spec/data-provider-widget.spec.js
@@ -4,63 +4,75 @@
  * www.laxarjs.org
  */
 define( [
+   'json!../widget.json',
+   'laxar-mocks',
    'angular-mocks',
-   '../data-provider-widget',
-   'laxar/laxar_testing',
    'laxar-patterns',
    './spec_data'
-], function( ngMocks, widgetModule, ax, patterns, specData ) {
+], function(descriptor, axMocks, ngMocks, patterns, specData ) {
    'use strict';
 
-   describe( 'A DataProviderWidget', function() {
+   describe( 'A data-provider-widget', function() {
+      var $httpBackend;
+      var $provide;
+      var widgetEventBus;
+      var widgetScope;
+      var testEventBus;
 
-      var testBed_;
+      ////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-      beforeEach( function setup() {
-         testBed_ = ax.testing.portalMocksAngular.createControllerTestBed( 'mashup-demo/data-provider-widget' );
-         testBed_.featuresMock = {
-            data: {
-               resource: 'timeSeriesData',
+      var widgetConfiguration = {
+         data: {
+            resource: 'timeSeriesData',
                items: [
-                  {
-                     title: 'Data-Set-1',
-                     location: 'data-set-1.json'
-                  },
-                  {
-                     title: 'Data-Set-2',
-                     location: 'data-set-2.json'
-                  },
-                  {
-                     title: 'Data-Set-Non-Existing',
-                     location: 'data-set-non-existing.json'
-                  }
-               ]
-            }
-         };
+               {
+                  title: 'Data-Set-1',
+                  location: 'data-set-1.json'
+               },
+               {
+                  title: 'Data-Set-2',
+                  location: 'data-set-2.json'
+               },
+               {
+                  title: 'Data-Set-Non-Existing',
+                  location: 'data-set-non-existing.json'
+               }
+            ]
+         }
+      };
 
-         testBed_.useWidgetJson();
-         testBed_.setup();
+      beforeEach( function() {
+         ngMocks.module( function( _$provide_ ) {
+            $provide = _$provide_;
+         } );
+      } );
+
+      beforeEach( axMocks.createSetupForWidget( descriptor ) );
+
+      beforeEach( function() {
+         axMocks.widget.configure( widgetConfiguration );
+      } );
+
+      beforeEach( axMocks.widget.load );
+
+      beforeEach( function() {
+         widgetScope = axMocks.widget.$scope;
+         widgetEventBus = axMocks.widget.axEventBus;
+         testEventBus = axMocks.eventBus;
+         ngMocks.inject( function( $injector ) {
+            $httpBackend = $injector.get( '$httpBackend' );
+         } );
       } );
 
       ////////////////////////////////////////////////////////////////////////////////////////////////////////
 
       afterEach( function() {
-         testBed_.tearDown();
+         axMocks.tearDown();
       } );
 
       ////////////////////////////////////////////////////////////////////////////////////////////////////////
 
       describe( 'with feature data', function() {
-
-         var $httpBackend;
-
-         beforeEach( function() {
-            ngMocks.inject( function( _$httpBackend_ ) {
-               $httpBackend = _$httpBackend_;
-            } );
-         } );
-
-         /////////////////////////////////////////////////////////////////////////////////////////////////////
 
          afterEach( function() {
             $httpBackend.verifyNoOutstandingExpectation();
@@ -72,11 +84,10 @@ define( [
          it( 'publishes a didReplace event with the file content of the new location when the corresponding button is pressed', function() {
             $httpBackend.expectGET( 'data-set-2.json' ).respond( specData.dataSet2 );
 
-            testBed_.scope.useItem( testBed_.featuresMock.data.items[1] );
-            jasmine.Clock.tick( 0 );
+            widgetScope.useItem (widgetConfiguration.data.items[ 1 ] );
             $httpBackend.flush();
 
-            expect( testBed_.scope.eventBus.publish ).toHaveBeenCalledWith( 'didReplace.timeSeriesData', {
+            expect( widgetEventBus.publish ).toHaveBeenCalledWith( 'didReplace.timeSeriesData', {
                   resource: 'timeSeriesData',
                   data: specData.dataSet2
                }, jasmine.any( Object )
@@ -88,14 +99,12 @@ define( [
          it( 'publishes a didEncounterError event if the new location is not available', function() {
             $httpBackend.expectGET( 'data-set-non-existing.json' ).respond( 404, 'Not Found' );
 
-            testBed_.scope.useItem( testBed_.featuresMock.data.items[2] );
-            jasmine.Clock.tick( 0 );
+            widgetScope.useItem (widgetConfiguration.data.items[ 2 ] );
             $httpBackend.flush();
 
-            expect( testBed_.scope.eventBus.publish ).toHaveBeenCalledWith( 'didEncounterError.HTTP_GET', jasmine.any( Object ), jasmine.any( Object ) );
+            expect( widgetEventBus.publish ).toHaveBeenCalledWith( 'didEncounterError.HTTP_GET', jasmine.any( Object ), jasmine.any( Object ) );
          } );
 
       } );
    } );
-
 } );