+++ includes/widgets/mashup-demo/table-editor-widget/spec/table-editor-widget.spec.js
@@ -4,33 +4,45 @@
  * www.laxarjs.org
  */
 define( [
-   '../table-editor-widget',
-   'laxar/laxar_testing',
+   'json!../widget.json',
+   'laxar-mocks',
+   'laxar',
    'laxar-patterns',
    './spec_data'
-], function( widgetModule, ax, patterns, specData ) {
+], function( descriptor, axMocks, ax, patterns, specData ) {
    'use strict';
 
    describe( 'A TableEditorWidget', function() {
+      var widgetEventBus;
+      var widgetScope;
+      var testEventBus;
 
-      var testBed_;
+      ////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-      beforeEach( function setup() {
-         testBed_ = ax.testing.portalMocksAngular.createControllerTestBed( 'mashup-demo/table-editor-widget' );
-         testBed_.featuresMock = {
+      beforeEach( axMocks.createSetupForWidget( descriptor, {
+         knownMissingResources: [ 'laxar-date-picker-control.css ' ]
+      } ) );
+
+      beforeEach( function() {
+         axMocks.widget.configure( {
             timeSeries: {
                resource: 'timeSeriesData'
             }
-         };
+         } );
+      } );
 
-         testBed_.useWidgetJson();
-         testBed_.setup();
+      beforeEach( axMocks.widget.load );
+
+      beforeEach( function() {
+         widgetScope = axMocks.widget.$scope;
+         widgetEventBus = axMocks.widget.axEventBus;
+         testEventBus = axMocks.eventBus;
       } );
 
       ////////////////////////////////////////////////////////////////////////////////////////////////////////
 
       afterEach( function() {
-         testBed_.tearDown();
+         axMocks.tearDown();
       } );
 
       ////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -39,40 +51,40 @@ define( [
 
          beforeEach( function() {
 
-            testBed_.eventBusMock.publish( 'didReplace.timeSeriesData', {
+            testEventBus.publish( 'didReplace.timeSeriesData', {
                resource: 'timeSeriesData',
                data: specData.originalResource
             } );
-            jasmine.Clock.tick( 0 );
+            testEventBus.flush();
          } );
 
          /////////////////////////////////////////////////////////////////////////////////////////////////////
 
          it( 'acts as a slave for the configured resource', function() {
-            expect( testBed_.scope.eventBus.subscribe )
+            expect( widgetEventBus.subscribe )
                .toHaveBeenCalledWith( 'didReplace.timeSeriesData', jasmine.any( Function ) );
-            expect( testBed_.scope.eventBus.subscribe )
+            expect( widgetEventBus.subscribe )
                .toHaveBeenCalledWith( 'didUpdate.timeSeriesData', jasmine.any( Function ) );
-            expect( testBed_.scope.resources.timeSeries ).toEqual( specData.originalResource );
+            expect( widgetScope.resources.timeSeries ).toEqual( specData.originalResource );
          } );
 
          /////////////////////////////////////////////////////////////////////////////////////////////////////
 
          it( 'shows the published resource as the data model of the table', function() {
-            expect( testBed_.scope.model.tableModel ).toEqual( specData.expectedTableModel );
+            expect( widgetScope.model.tableModel ).toEqual( specData.expectedTableModel );
          } );
 
          /////////////////////////////////////////////////////////////////////////////////////////////////////
 
          it( 'publishes a didUpdate event after the user changed a data value', function() {
-            testBed_.scope.model.tableModel[1][1] = 11;
-            testBed_.scope.$emit( 'axTableEditor.afterChange' );
+            widgetScope.model.tableModel[1][1] = 11;
+            widgetScope.$emit( 'axTableEditor.afterChange' );
 
             var expectedResource = ax.object.deepClone( specData.originalResource );
             expectedResource.series[ 0 ].values[ 0 ] = 11;
             var patch = patterns.json.createPatch( specData.originalResource, expectedResource );
 
-            expect( testBed_.scope.eventBus.publish )
+            expect( widgetEventBus.publish )
                .toHaveBeenCalledWith( 'didUpdate.timeSeriesData', {
                   resource: 'timeSeriesData',
                   patches: patch
@@ -82,13 +94,13 @@ define( [
          /////////////////////////////////////////////////////////////////////////////////////////////////////
 
          it( 'ignores rows on which the time grid tick is removed in the didUpdate event data', function() {
-            testBed_.scope.model.tableModel[2][0] = null;
-            testBed_.scope.$emit( 'axTableEditor.afterChange' );
+            widgetScope.model.tableModel[2][0] = null;
+            widgetScope.$emit( 'axTableEditor.afterChange' );
 
             var expectedResource = specData.expectedResourceWithRemovedTimeGridTick;
             var patch = patterns.json.createPatch( specData.originalResource, expectedResource );
 
-            expect( testBed_.scope.eventBus.publish )
+            expect( widgetEventBus.publish )
                .toHaveBeenCalledWith( 'didUpdate.timeSeriesData', {
                   resource: 'timeSeriesData',
                   patches: patch
@@ -98,13 +110,13 @@ define( [
          /////////////////////////////////////////////////////////////////////////////////////////////////////
 
          it( 'ignores columns on which the series label is removed in the didUpdate event data', function() {
-            testBed_.scope.model.tableModel[0][2] = null;
-            testBed_.scope.$emit( 'axTableEditor.afterChange' );
+            widgetScope.model.tableModel[0][2] = null;
+            widgetScope.$emit( 'axTableEditor.afterChange' );
 
             var expectedResource = specData.expectedResourceWithRemovedSeriesLabel;
             var patch = patterns.json.createPatch( specData.originalResource, expectedResource );
 
-            expect( testBed_.scope.eventBus.publish )
+            expect( widgetEventBus.publish )
                .toHaveBeenCalledWith( 'didUpdate.timeSeriesData', {
                   resource: 'timeSeriesData',
                   patches: patch
@@ -114,13 +126,13 @@ define( [
          /////////////////////////////////////////////////////////////////////////////////////////////////////
 
          it( 'replaces the resource and model after receiving a new resource', function() {
-            testBed_.eventBusMock.publish( 'didReplace.timeSeriesData', {
+            testEventBus.publish( 'didReplace.timeSeriesData', {
                resource: 'timeSeriesData',
                data: specData.otherResource
             } );
-            jasmine.Clock.tick( 0 );
-            expect( testBed_.scope.resources.timeSeries ).toEqual( specData.otherResource );
-            expect( testBed_.scope.model.tableModel ).toEqual( specData.expectedTableModelForOtherResource );
+            testEventBus.flush();
+            expect( widgetScope.resources.timeSeries ).toEqual( specData.otherResource );
+            expect( widgetScope.model.tableModel ).toEqual( specData.expectedTableModelForOtherResource );
          } );
       } );